+++ packages/resolver/lib/sources/truffle/index.ts
@@ -4,7 +4,6 @@ import { Deployed } from "./Deployed";
 import findContracts from "@truffle/contract-sources";
 import { ResolverSource } from "../../source";
 const contract = require("@truffle/contract");
-const {format} = require("web3-providers-http-proxy");
 export class Truffle implements ResolverSource {
   options: any;
 
@@ -51,7 +50,7 @@ export class Truffle implements ResolverSource {
       const addresses = filesData.map(data => {
         const c = contract(JSON.parse(data));
         c.setNetwork(this.options.network_id);
-        if (c.isDeployed()) return format.formatHexAddress(c.address);
+        if (c.isDeployed()) return c.address;
         return null;
       });